Changed `xfail` to `skip_translation` for `tests/asm.aarch64`.
`c2rust transpile` fails completely on aarch64 with clang 10 on any file,
so we want to skip translation for now.

`xfail` would be correct for clang >10,
but for some reason it still tries to compile the whole crate afterward,
which of course fails for the same reason.
I can't easily change the `test_translator.py` logic for this, though,
since other `xfail`s depend on it working this way it seems.