Fixed the `Vagrantfile` to read the toolchain channel correctly from `rust-toolchain.toml`.
I'm not sure about this code because I'm not sure how to test the `Vagrantfile`.
I also don't know Ruby, but I tested this snippet of code in a separate Ruby file and it works.
Also, I'm not sure if there's a simple way to get ruby gems installed for use in the `Vagrantfile`,
because then I could just add a real toml parser instead of rolling part of my own.